+++ src/lib/process/process.ts
@@ -1,4 +1,4 @@
-import { EventListener } from "./events";
+import { EventListener } from "../events";
 import Scheduler from "./scheduler";
 
 type ProcessScheduler = typeof Scheduler;
